+++ src/main/java/org/exampleHotel/domain/room/RoomRepository.java
@@ -8,9 +8,9 @@ public interface RoomRepository {
 
     void readAll();
 
-    void remove(int id);
+    void remove(long id);
 
-    void edit(int id, int number, List<BedType> bedTypes);
+    void edit(long id, int number, List<BedType> bedTypes);
 
     Room getById(long id);
