+++ src/main/java/org/exampleHotel/ui/gui/reservations/AddNewReservationScene.java
@@ -45,7 +45,7 @@ gridPane.setVgap(14);
 
             allAsDTO.forEach( dto -> {
                 roomSelectionItems.add(
-                        new RoomSelectionItem(dto.getNumber(), dto.getId()));
+                        new RoomSelectionItem(dto.getNumber(), (int) dto.getId()));
             });
 
             List<GuestSelectionItem> guestSelectionItems = new ArrayList<>();
